DB-10455 fix mem
