+++ hbase_sql/src/test/java/com/splicemachine/derby/impl/sql/execute/operations/ExternalTableIT.java
@@ -2722,4 +2722,25 @@ public class ExternalTableIT extends SpliceUnitTest {
             Assert.fail("can't list " + path);
         }
     }
+
+    @Test
+    public void testAnalyzeExternalTable() throws Exception {
+        String path = getResourceDirectory() + "parquet_simple_file_test";
+
+        try (ResultSet rs = methodWatcher.executeQuery("CALL SYSCS_UTIL.ANALYZE_EXTERNAL_TABLE('" + path + "')") ) {
+            StringBuilder sb = new StringBuilder();
+            while( rs.next() ) {
+                sb.append(rs.getString(1) + "\n");
+            }
+            String expected = "CREATE EXTERNAL TABLE T (\n" +
+                    " column1 CHAR/VARCHAR(x),\n" +
+                    " column2 CHAR/VARCHAR(x),\n" +
+                    " partition1 CHAR/VARCHAR(x) \n" +
+                    ") PARTITIONED BY(\n" +
+                    " partition1 \n" +
+                    ")\n" +
+                    " STORED AS PARQUET LOCATION '%s';\n";
+            Assert.assertEquals(String.format(expected, path),  sb.toString());
+        }
+    }
 }
