+++ splice_si_api/src/test/java/com/splicemachine/si/impl/TestingFileSystem.java
@@ -22,9 +22,12 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.file.*;
+import java.nio.file.attribute.PosixFilePermission;
+import java.nio.file.attribute.PosixFilePermissions;
 import java.nio.file.spi.FileSystemProvider;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 import java.util.regex.Pattern;
 
 /**
@@ -232,7 +235,7 @@ public class TestingFileSystem extends DistributedFileSystem{
 
         @Override
         public String getGroup(){
-            throw new UnsupportedOperationException("IMPLEMENT");
+            return "memGroup";
         }
 
         @Override
@@ -261,18 +264,27 @@ public class TestingFileSystem extends DistributedFileSystem{
 
         @Override
         public FileInfo[] listDir(){
-            throw new UnsupportedOperationException("IMPLEMENT");
+            try {
+                return Files.list(p).map( path -> new PathInfo(path) ).toArray(PathInfo[]::new);
+            } catch (IOException e) {
+                e.printStackTrace();
+                return null;
+            }
         }
 
         @Override
         public String getPermissionStr() {
-            return "";
+            try {
+                Set<PosixFilePermission> s = Files.getPosixFilePermissions(p, LinkOption.NOFOLLOW_LINKS);
+                return PosixFilePermissions.toString(s);
+            } catch (IOException e) {
+                return "err";
+            }
         }
 
         @Override
-        public long getModificationTime()
-        {
-            return 0;
+        public long getModificationTime() {
+            return p.toFile().lastModified();
         }
         @Override
         public long size(){
